[map] [MenuSingle] Instanciacion de mapas con diferentes avatares, colores y escenarios a partir de un map_pattern.txt templatizado.
Se han embebido las siguientes propiedades o "keywords", configuradas a través del formulario de Player, tanto en archetypes como en map_pattern:
- COLOR: se sustituye por el color del jugador (R, G, B, Y); si aplica sobre el atributo modelColorN, teñirá del color correspondiente la subentidad de orden N del modelo gráfico asociado a la entidad/arquetipo.
- AVATAR: establece el archetype del avatar a utilizar por la instancia Player patrón (en map_pattern)
- CONTROL: se sustituye por el type NPC o Player, dependiendo de si el archetype del avatar va a ser controlado por el jugador o por IA.
- LOCAL:  establece si la instancia del Player patrón va a ser un jugador local (true) o remoto/NPC (false)
- SCENE: establece el archetype del scene a utilizar por la instancia Scene patrón (map_pattern).
- NICK: se sustitye por el nickname escogido por el Player; determina el nombre de la entidad de la instancia Player.
CPlayerSettings: nueva estructura de información serializable utilizada para leer el formulario y generar las propiedades anteriores.

ATTENTION: se han eliminado los antiguos mapas específicos (mapBlue, mapRed, etc...). A partir de ahora sólo debemos utilizar el map_pattern como patrón de todos los mapas.
Esto asentaría las bases para poder completar (por imitación de MenuSingle) los Lobbies de red y después comenzar con el multijugador.

[media]
 Eliminados los marine materials y MiShader -> Los colores ahora se establecen mediante código  + Normal_AO_Specular_Mappin_Shader
Añadidos OrbCOLOR.hfx restantes -> PENDIENTE generarlos con el color correcto.

[arquitectura] Muchos pequeños cambios difíciles de describir.

#44058321
Finish #57288774

+++ Src/Application/TimeConstants.h
@@ -2,7 +2,7 @@
 #define __Time_Constants_H
 
 const unsigned int TIMESTEP_USECS = 20000;
-const unsigned int TIMESTEP_MSECS = TIMESTEP_USECS * 0.001f;
+const unsigned int TIMESTEP_MSECS = (const unsigned int) TIMESTEP_USECS * 0.001f;
 const float TIMESTEP_SECS = TIMESTEP_USECS * 0.000001f;
 
 #endif
